Updated multiple different parts of code. Not sure what all I have done since it has been a little bit. However most work recently has been in the InsultGenerator.cs file. I can now update the nouns and adjectives lists with the set file path.
Most recent problem: Could not find files to update the object list with.

Solution: Remove the asterisk (*) from file path string. The asterisk was not doing what I thought it was. Instead of "*/WordListFiles/blah.txt" I have corrected it to "WordListFiles/blah.txt" in order to find the proper files.