Not a whole lot to add in this update. Mostly has to do with adding more documentation. I went through the files and added XML comments to the function so that it is easier to see what is going on. I refrained from adding comments to the Program.cs file, since it is constantly changing.
I also did a bit of changing to the InsultGenerator.cs file. I added a function that allows you to generate an insult directed towards a particular person by passing their name as a parameter. Passing the discord user mention causes this to mention the user that is being insulted.