I think this should be the finishing touch for now!!!!
I wrote a function to easily register your discord and riot account into the database. It puts you in and updates your match history so that none of the cells are null.

A lot of work needs to be done to clean it up more likely than not. But it should be a working product now.